Split out updaters and constants
