+++ src/constants.js
@@ -1,0 +1,6 @@
+export const modes = {
+  normal: 'NORMAL',
+  presenter: 'PRESENTER',
+  overview: 'OVERVIEW',
+  grid: 'GRID',
+}
